[Refactor][BE] userBookmark 로직들 리팩터링 완료
