+++ src/main/java/kr/co/ohgoodfood/controller/users/UsersController.java
@@ -1,8 +1,6 @@
 package kr.co.ohgoodfood.controller.users;
 
-import java.util.Collections;
 import java.util.List;
-import java.util.Map;
 
 import javax.servlet.http.HttpSession;
 
@@ -30,15 +28,9 @@ import lombok.extern.slf4j.Slf4j;
  *
  * 사용자 페이지 전용 기능을 처리하는 컨트롤러입니다.
  * - POST /user/signup	              : 사용자 회원가입 페이지
- * - GET  /user/main                  : 사용자 메인 화면 조회
- * - POST /user/filter/store          : AJAX 기반 가게 목록 필터링
- * - GET  /user/bookmark              : 해당 user_id가 가진 bookmark 목록 조회
- * - POST /user/bookmark/delete       : 해당하는 bookmark 삭제
- * - POST /user/bookmark/insert       : 해당하는 bookmark 추가
  * - GET  /user/main/orderList        : 유저가 가진 orderList 목록 조회
  * - POST /user/filter/order          : AJAX 기반 오더 목록 필터링
  * - POST /user/order/cancel          : 유저가 선택한 오더 주문 취소
- * - GET  /user/map/pin               : AJAX 기반 핀으로 선택한 스토어 fragment 조회
  * - GET  /user/mypage                : 유저 mypage 이동
  * - GET  /user/reviewList            : 하단 메뉴바 Review탭 이동시 전체 리뷰 목록 조회
  *
@@ -55,79 +47,6 @@ public class UsersController {
     // 지도 사용을 위한 앱키
     @Value("${kakao.map.appKey}")
     private String kakaoMapAppKey;
-
-    /**
-     * 해당 user가 가진 북마크 리스트를 조회한다.
-     *
-     * @param userMainFilter 요청 파라미터와 바인딩되어 뷰로 전달되는 DTO
-     * @param model          뷰에 전달할 데이터(Model)
-     * @param session        현재 HTTP 세션(로그인된 사용자 정보)
-     * @return               users/userBookmark.jsp로 포워딩
-     */
-    @GetMapping("/bookmark")
-    public String userBookmark(@ModelAttribute UserMainFilter userMainFilter,
-                               Model model,
-                               HttpSession session){
-
-        //세션에서 받아오는 로직
-        Account loginUser = (Account) session.getAttribute("user");
-        String user_id = loginUser.getUser_id();
-
-        List<Bookmark> bookmarkList = usersService.getBookmarkList(user_id);
-        model.addAttribute("bookmarkList", bookmarkList);
-
-        return "users/userBookmark"; // /WEB-INF/views/users/userBookmark.jsp로 forwarding
-    }
-
-    /**
-     * 해당 user가 가진 북마크 리스트 중, 특정 북마크를 삭제한다.
-     *
-     * @param bookmarkFilter bookMark delete에 필요한 필드 정보가 담긴 DTO
-     * @param model          뷰에 전달할 데이터(Model)
-     * @param session        현재 HTTP 세션(로그인된 사용자 정보)
-     * @return               json 응답, 성공시 {"code" : 200} / 실패시 {"code" : 500}
-     */
-    @PostMapping("/bookmark/delete")
-    @ResponseBody //json으로 code응답을 주기 위함이다.
-    public Map<String,Integer> userBookmarkDelete(@RequestBody BookmarkFilter bookmarkFilter,
-                                                  Model model,
-                                                  HttpSession session){
-        //세션에서 받아오는 로직
-        Account loginUser = (Account) session.getAttribute("user");
-        String user_id = loginUser.getUser_id();
-
-        //bookmark를 위해 user_id 세팅
-        bookmarkFilter.setUser_id(user_id);
-
-        //delete bookmark 실행
-        boolean result = usersService.deleteUserBookMark(bookmarkFilter);
-        return Collections.singletonMap("code", result ? 200 : 500);
-    }
-
-    /**
-     * 해당 user가 가진 북마클 리스트에서 삭제 된 것을 살리기 위함이다.
-     *
-     * @param bookmarkFilter bookMark delete에 필요한 필드 정보가 담긴 DTO
-     * @param model          뷰에 전달할 데이터(Model)
-     * @param session        현재 HTTP 세션(로그인된 사용자 정보)
-     * @return               json 응답, 성공시 {"code" : 200} / 실패시 {"code" : 500}
-     */
-    @PostMapping("/bookmark/insert")
-    @ResponseBody //json으로 code응답을 주기 위함이다.
-    public Map<String,Integer> userBookmarkInsert(@RequestBody BookmarkFilter bookmarkFilter,
-                                                  Model model,
-                                                  HttpSession session){
-        //세션에서 받아오는 로직
-        Account loginUser = (Account) session.getAttribute("user");
-        String user_id = loginUser.getUser_id();
-
-        //bookmark를 위해 user_id 세팅
-        bookmarkFilter.setUser_id(user_id);
-
-        //delete bookmark 실행
-        boolean result = usersService.insertUserBookMark(bookmarkFilter);
-        return Collections.singletonMap("code", result ? 200 : 500);
-    }
 
     /**
      * 세션에 있는 유저가 가진 주문 목록을 조회한다.
@@ -203,23 +122,6 @@ public class UsersController {
         }
         return "redirect:/user/orderList";
     }
-
-//    /**
-//     * map에서 pin 선택한 가게의 정보를 AJAX로 조회하고 뷰 프래그먼트만 반환한다.
-//     *
-//     * @param userMainFilter JSON 바디로 전달된 필터 정보 (필터 DTO에 자동 매핑)
-//     * @return               가게 정보를 포함한 JSP 프래그먼트 ("users/fragment/userMapPinStore")
-//     */
-//    @GetMapping("/map/pin")
-//    public String getMapPinStore(@ModelAttribute UserMainFilter userMainFilter,
-//                                 Model model){
-//
-//        MainStore mainStore = usersService.getMainStoreOne(userMainFilter);
-//        model.addAttribute("mainStore", mainStore);
-//
-//        //fragment return
-//        return "users/fragment/userMapPinStore";
-//    }
 
     /**
      *  사용자 회원가입
