+++ src/main/java/kr/co/ohgoodfood/dao/UserMapper.java
@@ -17,34 +17,6 @@ import java.util.List;
 @Mapper
 public interface UserMapper {
     /**
-     * 사용자 북마크 화면에서 표시할 가게 목록을 조회
-     *
-     * @param user_id          조회 대상 user_id
-     * @return                 필터 적용된 MainStore 리스트
-     */
-    List<Bookmark> selectAllBookmark(String user_id);
-
-    /**
-     * 사용자의 특정 북마크를 삭제 처리
-     *
-     * @param user_id          조회 대상 user_id
-     * @param store_id         user_id + store_id 조합으로 삭제
-     * @return                 영향받은 행(row) 수
-     */
-    int deleteBookmark(@Param("user_id") String user_id,
-                       @Param("store_id") String store_id);
-
-    /**
-     * 사용자 북마크 추가
-     *
-     * @param user_id          조회 대상 user_id
-     * @param store_id         북마크에 추가할 store 정보
-     * @return                 영향받은 행(row) 수
-     */
-    int insertBookmark(@Param("user_id") String user_id,
-                       @Param("store_id") String store_id);
-
-    /**
      * 사용자의 모든 주문내역을 출력
      *
      * @param userOrderFilter  필터 DTO
