+++ src/main/resources/kr/co/ohgoodfood/dao/UserBookmarkMapper.xml
@@ -1,0 +1,86 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!-- [refactor] bookmark 위한 전용 mapper -->
+<!DOCTYPE mapper
+        PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
+        "http:/mybatis.org/dtd/mybatis-3-mapper.dtd">
+<mapper namespace="kr.co.ohgoodfood.dao.UserBookmarkMapper">
+    <resultMap id="BookmarkDTO" type="kr.co.ohgoodfood.dto.BookmarkDTO">
+    <!-- store 연관관계 매핑 -->
+        <association property="store" javaType="kr.co.ohgoodfood.dto.Store">
+            <id     column="store_id"          property="store_id"/>
+            <result column="store_name"        property="store_name"/>
+            <result column="store_status"      property="store_status"/>
+            <result column="store_menu"      property="store_menu"/>
+            <result column="category_bakery"   property="category_bakery"/>
+            <result column="category_fruit"   property="category_fruit"/>
+            <result column="category_salad"   property="category_salad"/>
+            <result column="category_others"   property="category_others"/>
+            <result column="closed_at"   property="closed_at"/>
+            <result column="confirmed"   property="confirmed"/>
+        </association>
+    <!-- product 연관관계 매핑 -->
+        <association property="product" javaType="kr.co.ohgoodfood.dto.Product">
+            <id     column="product_no"          property="product_no"/>
+            <result column="pickup_start"        property="pickup_start"/>
+            <result column="pickup_end"      property="pickup_end"/>
+            <result column="origin_price"   property="origin_price"/>
+            <result column="sale_price"   property="sale_price"/>
+            <result column="reservation_end"   property="reservation_end"/>
+            <result column="amount"   property="amount"/>
+            <result column="product_explain"   property="product_explain"/>
+        </association>
+    <!-- Image 연관관계 매핑 -->
+        <association property="image" javaType="kr.co.ohgoodfood.dto.Image">
+            <result column="store_img"   property="store_img"/>
+        </association>
+
+    <!-- Bookmark 연관관계 매핑 -->
+        <association property="bookmark" javaType="kr.co.ohgoodfood.dto.Bookmark">
+            <id    column="bookmark_no"   property="bookmark_no"/>
+            <result column="user_id"   property="user_id"/>
+        </association>
+    </resultMap>
+    <!-- selectAllBookmark mapper, bookmark에서 이용한다. -->
+    <select id="selectAllBookmark" parameterType="map" resultMap="BookmarkDTO">
+        SELECT
+            s.store_id AS store_id, s.store_name AS store_name, s.store_menu AS store_menu, s.store_status AS store_status, s.category_bakery AS category_bakery,s.category_fruit AS category_fruit, s.category_salad AS category_salad, s.category_others  AS category_others, s.closed_at AS closed_at,
+            pf.product_no AS product_no, pf.pickup_start AS pickup_start,pf.pickup_end AS pickup_end, pf.origin_price AS origin_price, pf.sale_price AS sale_price, pf.product_explain AS product_explain, pf.reservation_end AS reservation_end, pf.amount AS amount,
+            img.first_img AS store_img
+        FROM Bookmark b
+        JOIN Store s ON s.store_id = b.store_id
+        LEFT JOIN Product pf
+            ON pf.store_id = s.store_id
+            AND DATE(pf.reservation_end) = CURDATE()
+        LEFT JOIN (
+            SELECT
+                store_id,
+                MIN(store_img) AS first_img
+            FROM Image
+            GROUP BY store_id
+        ) img ON img.store_id = s.store_id
+        <where>
+            s.confirmed = 'Y'
+            AND b.user_id = #{user_id}
+            <!-- ORDER BY 순서 : 오픈, 매진, 마감순 정렬 -->
+        </where>
+        ORDER BY
+            CASE
+                WHEN s.store_status = 'Y' AND amount >  0 THEN 1
+                WHEN s.store_status = 'Y' AND amount =  0 THEN 2
+            ELSE 3
+        END;
+    </select>
+
+    <!-- deleteBookmark mapper -->
+    <delete id="deleteBookmark" parameterType="map">
+        DELETE FROM Bookmark
+            WHERE store_id = #{store_id}
+            AND user_id = #{user_id};
+    </delete>
+
+    <!-- insertBookmark mapper -->
+    <delete id="insertBookmark" parameterType="map">
+        INSERT INTO Bookmark (user_id, store_id)
+        VALUES (#{user_id}, #{store_id});
+    </delete>
+</mapper>
